Move Hough transform classes to segment.hough package.
First it is indeed a segmentation technique.
Second it matches @gselzer work in another branch.
