+++ src/main/java/net/imagej/ops/segment/hough/HoughCircleDetectorOp.java
@@ -1,4 +1,4 @@
-package net.imagej.ops.hough;
+package net.imagej.ops.segment.hough;
 
 import net.imagej.ops.Op;
